feat(template-x): separate search api
